bugfix: 게임 종료시 기존 UI 제거 & Roar 적중 시 쿨타임 감소 & Cursed Howl 사용 시 침식 즉시 적용
