+++ Client/Assets/Scripts/Contents/Skills/Roar.cs
@@ -34,7 +34,7 @@ public class Roar : BaseSkill
                             IsHit = true;
                             crew.Rpc_OnDamaged(SkillData.Damage);
                             crew.Rpc_OnSanityDamaged(SkillData.SanityDamage);
-                            Owner.SkillController.Skills[3].CurrentCoolTime -= 20f;
+                            Owner.SkillController.Skills[2].CurrentCoolTime -= 20f;
                         }
                     }
                 }
