+++ Client/Assets/Scripts/UI/Scene/UI_CrewIngame.cs
@@ -1,7 +1,4 @@
 using System;
-using System.Collections;
-using UnityEngine.UI;
-using TMPro;
 using UnityEngine;
 
 public class UI_CrewIngame : UI_Ingame
@@ -12,9 +9,6 @@ public class UI_CrewIngame : UI_Ingame
     public UI_CrewStamina CrewStaminaUI { get; private set; }
 
     public UI_CrewWin UICrewWin { get; private set; }
-
-    public Canvas Canvas;
-    public Camera camera;
 
     private Crew Crew {
         get => Creature as Crew;
@@ -63,23 +57,5 @@ public class UI_CrewIngame : UI_Ingame
         CrewHpUI.gameObject.SetActive(false);
         CrewStaminaUI.gameObject.SetActive(false);
         CurrentSectorUI.gameObject.SetActive(false);
-    }
-
-    public void EndGame()
-    {
-        Canvas.renderMode = RenderMode.ScreenSpaceCamera;
-        Transform[] children = gameObject.GetComponentsInChildren<Transform>();
-
-        // 자식 객체들 중에서 Camera 컴포넌트를 가진 객체 찾기
-        foreach (Transform child in children)
-        {
-            // 자식 객체가 Camera 컴포넌트를 가지고 있는지 확인
-            Camera childCamera = child.GetComponent<Camera>();
-            if (childCamera != null)
-            {
-                camera = childCamera;
-            }
-        }
-        Canvas.worldCamera = camera;
     }
 }
